backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
fastapi-cors
bytez
google-generativeai
